refactor(desktop): split updater support files
Problem:
The updater and updater tests exceeded the repository line-count budgets after the packaging compatibility work.

Root cause:
New bundle resolution, manifest asset definitions, and architecture-download tests were added to already oversized source files.

Fix:
Move the new responsibilities into focused support files and regenerate the desktop inventory without raising the lint baseline.

Verification:
go run ./tools/repolint
go test ./tools/desktopinventory
cd desktop && go test ./...
